smart_attendance_tracker_v2/requirements.txt
Flask==2.2.5
Flask-SQLAlchemy==3.0.3
python-dotenv==1.1.0
twilio==8.4.0
mysql-connector-python==8.1.0
